moved pose estimation adapters to separate modules
