+++ tools/accuracy_checker/accuracy_checker/adapters/__init__.py
@@ -63,13 +63,9 @@ from .retinanet import RetinaNetAdapter, MultiOutRetinaNet
 from .yolo import TinyYOLOv1Adapter, YoloV2Adapter, YoloV3Adapter, YoloV3ONNX
 from .classification import ClassificationAdapter
 from .segmentation import SegmentationAdapter, BrainTumorSegmentationAdapter, DUCSegmentationAdapter
-from .pose_estimation import (
-    OpenPoseAdapter,
-    AssociativeEmbeddingAdapter,
-    HumanPoseAdapter,
-    SingleHumanPoseAdapter,
-    StackedHourGlassNetworkAdapter
-)
+from .pose_estimation import HumanPoseAdapter, SingleHumanPoseAdapter, StackedHourGlassNetworkAdapter
+from .pose_estimation_openpose import OpenPoseAdapter
+from .pose_estimation_associative_embedding import AssociativeEmbeddingAdapter
 
 from .pose_estimation_3d import HumanPose3dAdapter
 
@@ -166,10 +162,10 @@ __all__ = [
     'LPRAdapter',
     'CTCGreedySearchDecoder',
 
-    'OpenPoseAdapter',
     'AssociativeEmbeddingAdapter',
     'HumanPoseAdapter',
     'HumanPose3dAdapter',
+    'OpenPoseAdapter',
     'SingleHumanPoseAdapter',
     'StackedHourGlassNetworkAdapter',
