+++ tools/accuracy_checker/accuracy_checker/adapters/pose_estimation.py
@@ -19,14 +19,12 @@ from operator import itemgetter
 
 import cv2
 import numpy as np
-from scipy.optimize import linear_sum_assignment
-from skimage.measure import block_reduce
 
 from ..adapters import Adapter
-from ..config import ConfigValidator, StringField, ConfigError, NumberField
+from ..config import ConfigValidator, StringField, ConfigError
 from ..preprocessor import ObjectCropWithScale
 from ..representation import PoseEstimationPrediction
-from ..utils import contains_all, contains_any
+from ..utils import contains_any
 
 
 class HumanPoseAdapter(Adapter):
@@ -372,408 +370,6 @@ class HumanPoseAdapter(Adapter):
         return persons_keypoints_x, persons_keypoints_y, persons_keypoints_v, scores
 
 
-class OpenPoseAdapter(Adapter):
-    __provider__ = 'human_pose_estimation_openpose'
-    prediction_types = (PoseEstimationPrediction, )
-
-    @classmethod
-    def parameters(cls):
-        parameters = super().parameters()
-        parameters.update({
-            'part_affinity_fields_out': StringField(
-                description="Name of output layer with keypoints pairwise relations (part affinity fields).",
-                optional=True
-            ),
-            'keypoints_heatmap_out': StringField(
-                description="Name of output layer with keypoints heatmaps.", optional=True
-            ),
-            'upscale_factor': NumberField(
-                description="Upscaling factor for output feature maps before postprocessing.",
-                value_type=float, min_value=1, default=1, optional=True
-            ),
-        })
-        return parameters
-
-    def validate_config(self):
-        super().validate_config(on_extra_argument=ConfigValidator.WARN_ON_EXTRA_ARGUMENT)
-
-    def configure(self):
-        self.upscale_factor = self.get_value_from_config('upscale_factor')
-        self.part_affinity_fields = self.get_value_from_config('part_affinity_fields_out')
-        self.keypoints_heatmap = self.get_value_from_config('keypoints_heatmap_out')
-        self.concat_out = self.part_affinity_fields is None and self.keypoints_heatmap is None
-        if not self.concat_out:
-            contains_both = self.part_affinity_fields is not None and self.keypoints_heatmap is not None
-            if not contains_both:
-                raise ConfigError(
-                    'human_pose_estimation adapter should contains both: keypoints_heatmap_out '
-                    'and part_affinity_fields_out or not contain them at all (in single output model case)'
-                )
-            self._keypoints_heatmap_bias = self.keypoints_heatmap + '/add_'
-            self._part_affinity_fields_bias = self.part_affinity_fields + '/add_'
-
-        self.decoder = OpenPoseDecoder(num_joints=18, delta=0.5 if self.upscale_factor == 1 else 0.0)
-        self.nms = HeatmapNMS(kernel=2 * int(np.round(6 / 7 * self.upscale_factor)) + 1)
-
-    def process(self, raw, identifiers, frame_meta):
-        result = []
-        raw_outputs = self._extract_predictions(raw, frame_meta)
-        if not self.concat_out:
-            if not contains_any(raw_outputs, [self.part_affinity_fields, self._part_affinity_fields_bias]):
-                raise ConfigError('part affinity fields output not found')
-            if not contains_any(raw_outputs, [self.keypoints_heatmap, self._keypoints_heatmap_bias]):
-                raise ConfigError('keypoints heatmap output not found')
-            keypoints_heatmap = raw_outputs[
-                self.keypoints_heatmap if self.keypoints_heatmap in raw_outputs else self._keypoints_heatmap_bias
-            ]
-            pafs = raw_outputs[
-                self.part_affinity_fields if self.part_affinity_fields in raw_outputs
-                else self._part_affinity_fields_bias
-            ]
-            raw_output = zip(identifiers, keypoints_heatmap, pafs, frame_meta)
-        else:
-            concat_out = raw_outputs[self.output_blob]
-            keypoints_num = concat_out.shape[1] // 3
-            keypoints_heat_map = concat_out[:, :keypoints_num, :]
-            pafs = concat_out[:, keypoints_num:, :]
-            raw_output = zip(identifiers, keypoints_heat_map, pafs, frame_meta)
-        for identifier, heatmap, paf, meta in raw_output:
-            output_h, output_w = heatmap.shape[-2:]
-            if self.upscale_factor > 1:
-                self.decoder.delta = 0
-                heatmap = np.transpose(heatmap, (1, 2, 0))
-                heatmap = cv2.resize(heatmap, (0, 0), fx=self.upscale_factor, fy=self.upscale_factor,
-                                     interpolation=cv2.INTER_CUBIC)
-                heatmap = np.transpose(heatmap, (2, 0, 1))
-                paf = np.transpose(np.squeeze(paf), (1, 2, 0))
-                paf = cv2.resize(paf, (0, 0), fx=self.upscale_factor, fy=self.upscale_factor,
-                                 interpolation=cv2.INTER_CUBIC)
-                paf = np.transpose(paf, (2, 0, 1))
-            hmap = heatmap[None]
-            nms_hmap = self.nms(hmap)
-            poses, scores = self.decoder(hmap, nms_hmap, paf[None])
-            if len(scores) == 0:
-                result.append(PoseEstimationPrediction(
-                    identifier,
-                    np.empty((0, 17), dtype=float),
-                    np.empty((0, 17), dtype=float),
-                    np.empty((0, 17), dtype=float),
-                    np.empty((0, ), dtype=float)
-                ))
-                continue
-            poses = poses.astype(float)
-            scores = np.asarray(scores).astype(float)
-            scale_x = meta['scale_x']
-            scale_y = meta['scale_y']
-            input_h, input_w = next(iter(meta['input_shape'].values()))[-2:]
-            output_scale_x = input_w / output_w
-            output_scale_y = input_h / output_h
-            poses[:, :, 0] *= output_scale_x / self.upscale_factor / scale_x
-            poses[:, :, 1] *= output_scale_y / self.upscale_factor / scale_y
-            point_scores = poses[:, :, 2]
-
-            result.append(PoseEstimationPrediction(
-                identifier,
-                poses[:, :, 0],
-                poses[:, :, 1],
-                point_scores,
-                scores))
-        return result
-
-
-class AssociativeEmbeddingDecoder:
-
-    def __init__(self, num_joints, max_num_people, detection_threshold, use_detection_val,
-                 ignore_too_much, tag_threshold,
-                 adjust=True, refine=True, delta=0.0, joints_order=None):
-        self.num_joints = num_joints
-        self.max_num_people = max_num_people
-        self.detection_threshold = detection_threshold
-        self.tag_threshold = tag_threshold
-        self.use_detection_val = use_detection_val
-        self.ignore_too_much = ignore_too_much
-
-        if self.num_joints == 17 and joints_order is None:
-            self.joint_order = (0, 1, 2, 3, 4, 5, 6, 11, 12, 7, 8, 9, 10, 13, 14, 15, 16)
-        else:
-            self.joint_order = list(np.arange(self.num_joints))
-
-        self.do_adjust = adjust
-        self.do_refine = refine
-        self.delta = delta
-
-    def match(self, tag_k, loc_k, val_k):
-        return list(map(self._match_by_tag, zip(tag_k, loc_k, val_k)))
-
-    @staticmethod
-    def _max_match(scores):
-        r, c = linear_sum_assignment(scores)
-        tmp = np.stack((r, c), axis=1)
-        return tmp
-
-    def _match_by_tag(self, inp):
-        tag_k, loc_k, val_k = inp
-
-        embd_size = tag_k.shape[2]
-
-        class Pose:
-            def __init__(self, num_joints, tag_size=1):
-                self.num_joints = num_joints
-                self.tag_size = tag_size
-                self.pose = np.zeros((num_joints, 2 + 1 + tag_size), dtype=np.float32)
-                self.pose_tag = np.zeros(tag_size, dtype=np.float32)
-                self.valid_points_num = 0
-
-            def add(self, idx, joint, tag):
-                self.pose[idx] = joint
-                self.pose_tag = (self.pose_tag * self.valid_points_num) + tag
-                self.valid_points_num += 1
-                self.pose_tag /= self.valid_points_num
-
-            @property
-            def tag(self):
-                if self.valid_points_num > 0:
-                    return self.pose_tag
-                return None
-
-        all_joints = np.concatenate((loc_k, val_k[..., None], tag_k), -1)
-
-        poses = []
-        for idx in self.joint_order:
-            tags = tag_k[idx]
-            joints = all_joints[idx]
-            mask = joints[:, 2] > self.detection_threshold
-            tags = tags[mask]
-            joints = joints[mask]
-
-            if joints.shape[0] == 0:
-                continue
-
-            if len(poses) == 0:
-                for tag, joint in zip(tags, joints):
-                    pose = Pose(self.num_joints, embd_size)
-                    pose.add(idx, joint, tag)
-                    poses.append(pose)
-            else:
-                if self.ignore_too_much and len(poses) == self.max_num_people:
-                    continue
-                poses_tags = np.stack([p.tag for p in poses], axis=0)
-
-                diff = tags[:, None] - poses_tags[None, :]
-                diff_normed = np.linalg.norm(diff, ord=2, axis=2)
-                diff_saved = np.copy(diff_normed)
-
-                if self.use_detection_val:
-                    diff_normed = np.round(diff_normed) * 100 - joints[:, 2:3]
-
-                num_added = diff.shape[0]
-                num_grouped = diff.shape[1]
-
-                if num_added > num_grouped:
-                    diff_normed = np.concatenate(
-                        (diff_normed, np.zeros((num_added, num_added - num_grouped), dtype=np.float32) + 1e10),
-                        axis=1)
-
-                pairs = self._max_match(diff_normed)
-                for row, col in pairs:
-                    if row < num_added and col < num_grouped and diff_saved[row][col] < self.tag_threshold:
-                        poses[col].add(idx, joints[row], tags[row])
-                    else:
-                        pose = Pose(self.num_joints, embd_size)
-                        pose.add(idx, joints[row], tags[row])
-                        poses.append(pose)
-
-        if len(poses) > 0:
-            ans = np.stack([p.pose for p in poses]).astype(np.float32)
-            tags = np.stack([p.tag for p in poses]).astype(np.float32)
-        else:
-            ans = np.empty((0, self.num_joints, 2 + 1 + embd_size), dtype=np.float32)
-            tags = np.empty((0, embd_size), dtype=np.float32)
-        return ans, tags
-
-    def top_k(self, heatmaps, tags):
-        N, K, H, W = heatmaps.shape
-        heatmaps = heatmaps.reshape(N, K, -1)
-        ind = heatmaps.argpartition(-self.max_num_people, axis=2)[:, :, -self.max_num_people:]
-        val_k = np.take_along_axis(heatmaps, ind, axis=2)
-        subind = np.argsort(-val_k, axis=2)
-        ind = np.take_along_axis(ind, subind, axis=2)
-        val_k = np.take_along_axis(val_k, subind, axis=2)
-
-        tags = tags.reshape(N, K, W * H, -1)
-        tag_k = [np.take_along_axis(tags[..., i], ind, axis=2) for i in range(tags.shape[3])]
-        tag_k = np.stack(tag_k, axis=3)
-
-        x = ind % W
-        y = ind // W
-        ind_k = np.stack((x, y), axis=3)
-
-        ans = {'tag_k': tag_k, 'loc_k': ind_k, 'val_k': val_k}
-        return ans
-
-    @staticmethod
-    def adjust(ans, heatmaps):
-        H, W = heatmaps.shape[-2:]
-        for n, people in enumerate(ans):
-            for person in people:
-                for k, joint in enumerate(person):
-                    heatmap = heatmaps[n, k]
-                    px = int(joint[0])
-                    py = int(joint[1])
-                    if 1 < px < W - 1 and 1 < py < H - 1:
-                        diff = np.array([
-                            heatmap[py, px + 1] - heatmap[py, px - 1],
-                            heatmap[py + 1, px] - heatmap[py - 1, px]
-                        ])
-                        joint[:2] += np.sign(diff) * .25
-        return ans
-
-    @staticmethod
-    def refine(heatmap, tag, keypoints, pose_tag=None):
-        K, H, W = heatmap.shape
-        if len(tag.shape) == 3:
-            tag = tag[..., None]
-
-        if pose_tag is not None:
-            prev_tag = pose_tag
-        else:
-            tags = []
-            for i in range(K):
-                if keypoints[i, 2] > 0:
-                    x, y = keypoints[i][:2].astype(int)
-                    tags.append(tag[i, y, x])
-            prev_tag = np.mean(tags, axis=0)
-
-        # Allocate the buffer for tags similarity matrix.
-        tag_copy = np.empty_like(tag[0, ..., 0])
-        for i, (_heatmap, _tag) in enumerate(zip(heatmap, tag)):
-            if keypoints[i, 2] > 0:
-                continue
-            tag_copy[...] = _tag[..., 0]
-            diff = tag_copy
-            diff -= prev_tag
-            np.abs(diff, out=diff)
-            np.floor(diff + 0.5, out=diff)
-            diff -= _heatmap
-            idx = diff.argmin()
-            y, x = np.divmod(idx, _heatmap.shape[-1])
-
-            # detection score at maximum position
-            val = _heatmap[y, x]
-
-            if val > 0:
-                keypoints[i, :3] = x, y, val
-                if 1 < x < W - 1 and 1 < y < H - 1:
-                    diff = np.array([
-                        _heatmap[y, x + 1] - _heatmap[y, x - 1],
-                        _heatmap[y + 1, x] - _heatmap[y - 1, x]
-                    ])
-                    keypoints[i, :2] += np.sign(diff) * .25
-
-        return keypoints
-
-    def __call__(self, heatmaps, tags, nms_heatmaps=None):
-        ans = self.match(**self.top_k(nms_heatmaps, tags))
-        ans, ans_tags = map(list, zip(*ans))
-
-        if self.do_adjust:
-            ans = self.adjust(ans, heatmaps)
-
-        if self.delta != 0.0:
-            for people in ans:
-                for person in people:
-                    for joint in person:
-                        joint[:2] += self.delta
-
-        ans = ans[0]
-        scores = np.asarray([i[:, 2].mean() for i in ans])
-
-        if self.do_refine:
-            heatmap_numpy = heatmaps[0]
-            tag_numpy = tags[0]
-            for i in range(len(ans)):
-                ans[i] = self.refine(heatmap_numpy, tag_numpy, ans[i], ans_tags[0][i])
-
-        return ans, scores
-
-
-class AssociativeEmbeddingAdapter(Adapter):
-    __provider__ = 'human_pose_estimation_ae'
-    prediction_types = (PoseEstimationPrediction, )
-
-    @classmethod
-    def parameters(cls):
-        parameters = super().parameters()
-        parameters.update({
-            'heatmaps_out': StringField(
-                description="Name of output layer with keypoints heatmaps.",
-                optional=True
-            ),
-            'nms_heatmaps_out': StringField(
-                description="Name of output layer with keypoints heatmaps after NMS.",
-                optional=True
-            ),
-            'embeddings_out': StringField(
-                description="Name of output layer with associative embeddings.",
-                optional=True
-            ),
-        })
-        return parameters
-
-    def validate_config(self):
-        super().validate_config(on_extra_argument=ConfigValidator.WARN_ON_EXTRA_ARGUMENT)
-
-    def configure(self):
-        self.heatmaps = self.get_value_from_config('heatmaps_out')
-        self.nms_heatmaps = self.get_value_from_config('nms_heatmaps_out')
-        self.embeddings = self.get_value_from_config('embeddings_out')
-        self.decoder = AssociativeEmbeddingDecoder(
-            num_joints=17,
-            adjust=True,
-            refine=True,
-            delta=0.0,
-            max_num_people=30,
-            detection_threshold=0.1,
-            tag_threshold=1,
-            use_detection_val=True,
-            ignore_too_much=False)
-
-    def process(self, raw, identifiers, frame_meta):
-        result = []
-        raw_outputs = self._extract_predictions(raw, frame_meta)
-        if not contains_all(raw_outputs, (self.heatmaps, self.nms_heatmaps, self.embeddings)):
-            raise ConfigError('Some of the outputs are not found')
-        raw_output = zip(identifiers, raw_outputs[self.heatmaps][None],
-                         raw_outputs[self.nms_heatmaps][None],
-                         raw_outputs[self.embeddings][None], frame_meta)
-
-        for identifier, heatmap, nms_heatmap, embedding, meta in raw_output:
-            poses, scores = self.decoder(heatmap, embedding, nms_heatmaps=nms_heatmap)
-            if len(scores) == 0:
-                result.append(PoseEstimationPrediction(
-                    identifier,
-                    np.empty((0, 17), dtype=float),
-                    np.empty((0, 17), dtype=float),
-                    np.empty((0, 17), dtype=float),
-                    np.empty((0, ), dtype=float)
-                ))
-                continue
-            poses = poses.astype(float)
-            scores = np.asarray(scores).astype(float)
-            scale_x = meta['scale_x']
-            scale_y = meta['scale_y']
-            poses[:, :, 0] /= scale_x / 2
-            poses[:, :, 1] /= scale_y / 2
-            point_scores = poses[:, :, 2]
-            result.append(PoseEstimationPrediction(
-                identifier,
-                poses[:, :, 0],
-                poses[:, :, 1],
-                point_scores,
-                scores))
-        return result
-
-
 class SingleHumanPoseAdapter(Adapter):
     __provider__ = 'single_human_pose_estimation'
     prediction_types = (PoseEstimationPrediction, )
@@ -900,295 +496,3 @@ class StackedHourGlassNetworkAdapter(Adapter):
             preds = preds.reshape(1, preds.size)
 
         return preds
-
-
-class HeatmapNMS:
-    def __init__(self, kernel):
-        self.kernel = kernel
-        self.pad = (kernel - 1) // 2
-
-    def max_pool(self, x):
-        # Max pooling kernel x kernel with stride 1 x 1.
-        k = self.kernel
-        p = self.pad
-        pooled = np.zeros_like(x)
-        hmap = np.pad(x, ((0, 0), (0, 0), (p, p), (p, p)))
-        h, w = hmap.shape[-2:]
-        for i in range(k):
-            n = (h - i) // k * k
-            for j in range(k):
-                m = (w - j) // k * k
-                hmap_slice = hmap[..., i:i + n, j:j + m]
-                pooled[..., i::k, j::k] = block_reduce(hmap_slice, (1, 1, k, k), np.max)
-        return pooled
-
-    def __call__(self, heatmaps):
-        pooled = self.max_pool(heatmaps)
-        return heatmaps * (pooled == heatmaps).astype(heatmaps.dtype)
-
-
-
-class OpenPoseDecoder:
-
-    BODY_PARTS_KPT_IDS = ((1, 2), (1, 5), (2, 3), (3, 4), (5, 6), (6, 7), (1, 8), (8, 9), (9, 10), (1, 11),
-                          (11, 12), (12, 13), (1, 0), (0, 14), (14, 16), (0, 15), (15, 17), (2, 16), (5, 17))
-    BODY_PARTS_PAF_IDS = (12, 20, 14, 16, 22, 24, 0, 2, 4, 6, 8, 10, 28, 30, 34, 32, 36, 18, 26)
-
-    def __init__(self, num_joints=18, skeleton=BODY_PARTS_KPT_IDS, paf_indices=BODY_PARTS_PAF_IDS,
-                 max_points=100, score_threshold=0.1, min_paf_alignment_score=0.05, delta=0.5):
-        self.num_joints = num_joints
-        self.skeleton = skeleton
-        self.paf_indices = paf_indices
-        self.max_points = max_points
-        self.score_threshold = score_threshold
-        self.min_paf_alignment_score = min_paf_alignment_score
-        self.delta = delta
-
-        self.points_per_limb = 10
-        self.grid = np.arange(self.points_per_limb, dtype=np.float32).reshape(1, -1, 1)
-
-    def __call__(self, heatmaps, nms_heatmaps, pafs):
-        batch_size, _, h, w = heatmaps.shape
-        assert batch_size == 1, 'Batch size of 1 only supported'
-
-        keypoints = self.extract_points(heatmaps, nms_heatmaps)
-        pafs = np.transpose(pafs, (0, 2, 3, 1))
-
-        if self.delta > 0:
-            for kpts in keypoints:
-                kpts[:, :2] += self.delta
-                np.core.umath.clip(kpts[:, 0], 0, w - 1, out=kpts[:, 0])
-                np.core.umath.clip(kpts[:, 1], 0, h - 1, out=kpts[:, 1])
-
-        pose_entries, keypoints = self.group_keypoints(keypoints, pafs, pose_entry_size=self.num_joints + 2)
-        poses, scores = self.convert_to_coco_format(pose_entries, keypoints)
-        if len(poses) > 0:
-            poses = np.asarray(poses, dtype=np.float32)
-            poses = poses.reshape((poses.shape[0], -1, 3))
-        else:
-            poses = np.empty((0, 17, 3), dtype=np.float32)
-            scores = np.empty(0, dtype=np.float32)
-
-        return poses, scores
-
-    def extract_points(self, heatmaps, nms_heatmaps):
-        batch_size, channels_num, h, w = heatmaps.shape
-        assert batch_size == 1, 'Batch size of 1 only supported'
-        assert channels_num >= self.num_joints
-
-        xs, ys, scores = self.top_k(nms_heatmaps)
-        masks = scores > self.score_threshold
-        all_keypoints = []
-        keypoint_id = 0
-        for k in range(self.num_joints):
-            # Filter low-score points.
-            mask = masks[0, k]
-            x = xs[0, k][mask].ravel()
-            y = ys[0, k][mask].ravel()
-            score = scores[0, k][mask].ravel()
-            n = len(x)
-            if n == 0:
-                all_keypoints.append(np.empty((0, 4), dtype=np.float32))
-                continue
-            # Apply quarter offset to improve localization accuracy.
-            x, y = self.refine(heatmaps[0, k], x, y)
-            np.core.umath.clip(x, 0, w - 1, out=x)
-            np.core.umath.clip(y, 0, h - 1, out=y)
-            # Pack resulting points.
-            keypoints = np.empty((n, 4), dtype=np.float32)
-            keypoints[:, 0] = x
-            keypoints[:, 1] = y
-            keypoints[:, 2] = score
-            keypoints[:, 3] = np.arange(keypoint_id, keypoint_id + n)
-            keypoint_id += n
-            all_keypoints.append(keypoints)
-        return all_keypoints
-
-    def top_k(self, heatmaps):
-        N, K, _, W = heatmaps.shape
-        heatmaps = heatmaps.reshape(N, K, -1)
-        # Get positions with top scores.
-        ind = heatmaps.argpartition(-self.max_points, axis=2)[:, :, -self.max_points:]
-        scores = np.take_along_axis(heatmaps, ind, axis=2)
-        # Keep top scores sorted.
-        subind = np.argsort(-scores, axis=2)
-        ind = np.take_along_axis(ind, subind, axis=2)
-        scores = np.take_along_axis(scores, subind, axis=2)
-        y, x = np.divmod(ind, W)
-        return x, y, scores
-
-    @staticmethod
-    def refine(heatmap, x, y):
-        h, w = heatmap.shape[-2:]
-        valid = np.logical_and(np.logical_and(x > 0, x < w - 1), np.logical_and(y > 0, y < h - 1))
-        xx = x[valid]
-        yy = y[valid]
-        dx = np.sign(heatmap[yy, xx + 1] - heatmap[yy, xx - 1], dtype=np.float32) * 0.25
-        dy = np.sign(heatmap[yy + 1, xx] - heatmap[yy - 1, xx], dtype=np.float32) * 0.25
-        x = x.astype(np.float32)
-        y = y.astype(np.float32)
-        x[valid] += dx
-        y[valid] += dy
-        return x, y
-
-    @staticmethod
-    def update_poses(part_id, kpt_a_id, kpt_b_id, all_keypoints, connections, pose_entries, pose_entry_size):
-        if part_id == 0:
-            pose_entries = [np.full(pose_entry_size, -1, dtype=np.float32) for _ in range(len(connections))]
-            for pose, connection in zip(pose_entries, connections):
-                pose[kpt_a_id] = connection[0]
-                pose[kpt_b_id] = connection[1]
-                pose[-1] = 2
-                # pose score = sum of all points' scores + sum of all connections' scores
-                pose[-2] = np.sum(all_keypoints[connection[0:2], 2]) + connection[2]
-        else:
-            for connection in connections:
-                pose_a_idx = -1
-                pose_b_idx = -1
-                for j, pose in enumerate(pose_entries):
-                    if pose[kpt_a_id] == connection[0]:
-                        pose_a_idx = j
-                    if pose[kpt_b_id] == connection[1]:
-                        pose_b_idx = j
-                if pose_a_idx < 0 and pose_b_idx < 0:
-                    # Create new pose entry.
-                    pose_entry = np.full(pose_entry_size, -1)
-                    pose_entry[kpt_a_id] = connection[0]
-                    pose_entry[kpt_b_id] = connection[1]
-                    pose_entry[-1] = 2
-                    pose_entry[-2] = np.sum(all_keypoints[connection[0:2], 2]) + connection[2]
-                    pose_entries.append(pose_entry)
-                elif pose_a_idx >= 0 and pose_b_idx >= 0:
-                    if pose_a_idx != pose_b_idx:
-                        # Merge two disjoint components into one pose.
-                        pose_a = pose_entries[pose_a_idx]
-                        pose_b = pose_entries[pose_b_idx]
-                        do_merge_poses = True
-                        for j in range(len(pose_b) - 2):
-                            if pose_a[j] >= 0 and pose_b[j] >= 0 and pose_a[j] != pose_b[j]:
-                                do_merge_poses = False
-                                break
-                        if not do_merge_poses:
-                            continue
-                        for j in range(len(pose_b) - 2):
-                            if pose_b[j] >= 0:
-                                pose_a[j] = pose_b[j]
-                        pose_a[-1] += pose_b[-1]
-                        pose_a[-2] += pose_b[-2] + connection[2]
-                        del pose_entries[pose_b_idx]
-                    else:
-                        # Adjust score of a pose.
-                        pose_entries[pose_a_idx][-2] += connection[2]
-                elif pose_a_idx >= 0:
-                    # Add a new limb into pose.
-                    pose = pose_entries[pose_a_idx]
-                    if pose[kpt_b_id] < 0:
-                        pose[-2] += all_keypoints[connection[1], 2]
-                    pose[kpt_b_id] = connection[1]
-                    pose[-2] += connection[2]
-                    pose[-1] += 1
-                elif pose_b_idx >= 0:
-                    # Add a new limb into pose.
-                    pose = pose_entries[pose_b_idx]
-                    if pose[kpt_a_id] < 0:
-                        pose[-2] += all_keypoints[connection[0], 2]
-                    pose[kpt_a_id] = connection[0]
-                    pose[-2] += connection[2]
-                    pose[-1] += 1
-        return pose_entries
-
-    def group_keypoints(self, all_keypoints_by_type, pafs, pose_entry_size=20):
-        all_keypoints = np.concatenate(all_keypoints_by_type, axis=0)
-        pose_entries = []
-        # For every limb.
-        for part_id, paf_channel in enumerate(self.paf_indices):
-            kpt_a_id, kpt_b_id = self.skeleton[part_id]
-            kpts_a = all_keypoints_by_type[kpt_a_id]
-            kpts_b = all_keypoints_by_type[kpt_b_id]
-            n = len(kpts_a)
-            m = len(kpts_b)
-            if n == 0 or m == 0:
-                continue
-
-            # Get vectors between all pairs of keypoints, i.e. candidate limb vectors.
-            a = kpts_a[:, :2]
-            a = np.broadcast_to(a[None], (m, n, 2))
-            b = kpts_b[:, :2]
-            vec_raw = (b[:, None, :] - a).reshape(-1, 1, 2)
-
-            # Sample points along every candidate limb vector.
-            steps = (1 / (self.points_per_limb - 1) * vec_raw)
-            points = steps * self.grid + a.reshape(-1, 1, 2)
-            points = points.round().astype(dtype=np.int32)
-            x = points[..., 0].ravel()
-            y = points[..., 1].ravel()
-
-            # Compute affinity score between candidate limb vectors and part affinity field.
-            part_pafs = pafs[0, :, :, paf_channel:paf_channel + 2]
-            field = part_pafs[y, x].reshape(-1, self.points_per_limb, 2)
-            vec_norm = np.linalg.norm(vec_raw, ord=2, axis=-1, keepdims=True)
-            vec = vec_raw / (vec_norm + 1e-6)
-            affinity_scores = (field * vec).sum(-1).reshape(-1, self.points_per_limb)
-            valid_affinity_scores = affinity_scores > self.min_paf_alignment_score
-            valid_num = valid_affinity_scores.sum(1)
-            affinity_scores = (affinity_scores * valid_affinity_scores).sum(1) / (valid_num + 1e-6)
-            success_ratio = valid_num / self.points_per_limb
-
-            # Get a list of limbs according to the obtained affinity score.
-            valid_limbs = np.where(np.logical_and(affinity_scores > 0, success_ratio > 0.8))[0]
-            affinity_scores = affinity_scores[valid_limbs]
-            b_idx, a_idx = np.divmod(valid_limbs, n)
-            if len(affinity_scores) == 0:
-                continue
-
-            # From all retrieved connections that share starting/ending keypoints leave only the top-scoring ones.
-            order = affinity_scores.argsort()[::-1]
-            affinity_scores = affinity_scores[order]
-            a_idx = a_idx[order]
-            b_idx = b_idx[order]
-            a_idx_unique = np.unique(a_idx, return_index=True)[1]
-            b_idx_unique = np.unique(b_idx, return_index=True)[1]
-            idx = np.intersect1d(a_idx_unique, b_idx_unique, assume_unique=True)
-            a = kpts_a[a_idx[idx], 3].astype(np.int32)
-            b = kpts_b[b_idx[idx], 3].astype(np.int32)
-            connections = list(zip(a, b, affinity_scores[idx]))
-
-            if len(connections) == 0:
-                continue
-
-            # Update poses with new connections.
-            pose_entries = self.update_poses(part_id, kpt_a_id, kpt_b_id, all_keypoints,
-                                             connections, pose_entries, pose_entry_size)
-
-        # Remove poses with not enough points.
-        filtered_entries = []
-        for pose in pose_entries:
-            if pose[-1] < 3:
-                continue
-            filtered_entries.append(pose)
-        pose_entries = np.asarray(filtered_entries, dtype=np.float32)
-        return pose_entries, all_keypoints
-
-    @staticmethod
-    def convert_to_coco_format(pose_entries, all_keypoints):
-        num_joints = 17
-        coco_keypoints = []
-        scores = []
-        for pose in pose_entries:
-            if len(pose) == 0:
-                continue
-            keypoints = np.zeros(num_joints * 3)
-            reorder_map = [0, -1, 6, 8, 10, 5, 7, 9, 12, 14, 16, 11, 13, 15, 2, 1, 4, 3]
-            person_score = pose[-2]
-            for keypoint_id, target_id in zip(pose[:-2], reorder_map):
-                if target_id < 0:
-                    continue
-                cx, cy, score = 0, 0, 0  # keypoint not found
-                if keypoint_id != -1:
-                    cx, cy, score = all_keypoints[int(keypoint_id), 0:3]
-                keypoints[target_id * 3 + 0] = cx
-                keypoints[target_id * 3 + 1] = cy
-                keypoints[target_id * 3 + 2] = score
-            coco_keypoints.append(keypoints)
-            scores.append(person_score * max(0, (pose[-1] - 1)))  # -1 for 'neck'
-        return np.asarray(coco_keypoints), np.asarray(scores)
